Implement sign ups
This commit sets up all the boilerplate and scaffolding required to
implement sign up as well.

Passwords are stored using the argon2id algorithm of the `buddy-hashers`
library.

We create middleware to wrap exceptions and keywordize form parameters.

We also write specs that are used for validation of form and user data.

We use Ring's flash middleware to inform the user if the sign up failed.