config: simplify configs and abolish env vars
We stop using a checked-in secrets file that reads environment
variables, as that is the worst of both worlds. Let's switch to a
simpler config, that assumes the secret is in a `secrets.edn` file as
before, but we won't read or rely on environment variables to manage
secrets.

This will also mean that our planned docker-compose deployment is able
to support '_FILE' based variables to get access to secrets.